created card component, made join us page as a temp
